+++ src/pages/index.ts
@@ -1,6 +1,6 @@
 import { LobbyPage } from "./LobbyPage";
 import { MeetingPage } from "./MeetingPage";
-import { StartPage } from "./StartPage";
+import { StartPage } from "./start-page/StartPage";
 import { LoginPage } from "./LoginPage";
 
 export { LobbyPage, MeetingPage, StartPage, LoginPage };
