+++ src/pages/start-page/components/StartButtons.tsx
@@ -1,0 +1,14 @@
+import { Button } from "../../../components/Button";
+
+const StartButtons = () => (
+  <div className="start-page_buttons">
+    <Button text="New meeting" classNames="meeting-button" onClick={() => {}} />
+    <Button
+      text="Join with a code"
+      classNames="join-button"
+      onClick={() => {}}
+    />
+  </div>
+);
+
+export { StartButtons };
