+++ src/styles/components/_all.scss
@@ -4,3 +4,4 @@
 @import "navbar";
 @import "menu";
 @import "button";
+@import "start-buttons";
