+++ src/styles/components/start-buttons.scss
@@ -1,0 +1,23 @@
+.start-page_buttons {
+  width: 100%;
+  display: flex;
+  gap: 1rem;
+
+  .meeting-button {
+    background-color: rgb(124, 188, 231);
+    width: 100%;
+    border-radius: 3px;
+    height: 2rem;
+    border: 1px solid hsl(234, 11%, 52%);
+    color: hsl(235, 19%, 35%);
+  }
+
+  .join-button {
+    background-color: transparent;
+    width: 100%;
+    border-radius: 3px;
+    height: 2rem;
+    border: 1px solid hsl(234, 11%, 52%);
+    color: rgb(124, 188, 231);
+  }
+}
