+++ src/styles/pages/start-page.scss
@@ -2,29 +2,23 @@
   position: relative;
   display: flex;
   align-items: center;
-  //justify-content: center;
   flex-direction: column;
   width: 100%;
   height: 100%;
   row-gap: 1rem;
   padding-right: 1rem;
   padding-left: 1rem;
-
-  .meeting-buttons {
-    display: flex;
-    width: 100%;
-    margin-top: 4.5rem;
-  }
+  padding-top: 5rem;
 
   .meetings {
     display: flex;
+    flex-direction: column;
     width: 100%;
     text-align: start;
+
     .title {
-      h5 {
-        color: hsl(236, 9%, 61%);
-        font-weight: lighter;
-      }
+      color: hsl(236, 9%, 61%);
+      font-weight: lighter;
     }
   }
 }
